Improve VM hardware fallback and runtime diagnostics
- Increase the DMA buffer registry capacity from 128 to 256 entries.
- Improve e1000 initialization for VM environments, including link setup,
  MMIO failures, unsupported hardware, and unavailable links.
- Avoid registering unusable Ethernet devices and skip DHCP when eth0 is
  missing or disconnected.
- Report unavailable USB and Wi-Fi hardware without treating it as a failure.
- Preserve network status output for registered interfaces without active links.
- Use explicit QEMU e1000 user-mode NAT networking.
- Fix framebuffer pitch validation for 64-bit row sizes.
- Run kernel self-tests after process manager initialization.
- Treat unavailable scheduler infrastructure as a skipped test.
- Prevent timer refresh and spike sampling checks from underflowing when the
  current tick precedes the next scheduled sample.

+++ kernel/drivers/video/fb.c
@@ -176,7 +176,7 @@ u8 fb_validate_geometry(u32 width, u32 height, u32 pitch, u32 bpp, u64 *bytes_ou
 
     row_bytes = (u64)width * ((u64)bpp / 8u);
     bytes = (u64)pitch * (u64)height;
-    if (row_bytes == 0u || bytes == 0u || pitch < (u32)row_bytes) {
+    if (row_bytes == 0u || bytes == 0u || (u64)pitch < row_bytes) {
         return 0;
     }
     if (bytes > UINT64_MAX - 4096ull) {
